feat: allow fractional bot poll interval for load testing

+++ bots/run.py
@@ -11,8 +11,9 @@ async def _run_bots(coroutines):
     await asyncio.gather(*coroutines)
 
 
+
 @app.command("start-bots")
-def start(count: int, strategy: str, interval: int = 1, starting_balance: float = 1000.00, base_api_url: str = "http://127.0.0.1:8000"):
+def start(count: int, strategy: str, interval: float = 1.0, starting_balance: float = 1000.00, base_api_url: str = "http://127.0.0.1:8000"):
 
     bot_instances = []
